Refactor trollflow2 to move main content outside of __init__.py
