+++ trollflow2/tests/__init__.py
@@ -1,8 +1,26 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2019 Pytroll developers
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>
+
 import sys
 import logging
 import unittest
 
-from trollflow2.tests import (test_trollflow2, test_launcher)
+from trollflow2.tests import (test_trollflow2, test_launcher, test_dict_tools)
 
 
 def suite():
@@ -13,6 +31,7 @@ def suite():
     mysuite = unittest.TestSuite()
     mysuite.addTests(test_trollflow2.suite())
     mysuite.addTests(test_launcher.suite())
+    mysuite.addTests(test_dict_tools.suite())
 
     return mysuite
